Finalize class roles
Variables are no longer stored inside manager classes
Transfer all loadAptDate methods to manager classes
Transfer saveEvent/Task methods from their manager classes to their base classes
