+++ app/src/main/java/com/fss/everythingapp/calendar/MonthViewController.java
@@ -74,7 +74,7 @@ public class MonthViewController {
             icon.setStrokeWidth(2.0);
             HBox.setMargin(icon, new Insets(1, 5, 1, 5));
 
-            DateManager loadedDate = dateLoader.getDateList().get(i);
+            Date loadedDate = dateLoader.getDateList().get(i);
 
             Label dateNameLabel = new Label();
             dateNameLabel.setPrefWidth(175);
@@ -115,7 +115,7 @@ public class MonthViewController {
             e.printStackTrace();
             paramDate = LocalDate.now();
         }
-        Date dateLoader = new Date('M', paramDate);
+        DateManager dateLoader = new DateManager('M', paramDate);
 
         for (int i = 0; i < dateLoader.getDateList().size(); i++) {
             Date loadedDate = dateLoader.getDateList().get(i);
